Sketcher: Improved ViewProvider visualisation for non-rational Bsplines with non-unitary weight constraints
===========================================================================================================

OCCT forces weights of non-rationa Bsplines to be 1.0

In the presence of non-unitary weight constraints, the weight representation (constraint) has a different radius
than the circle.

This commit fixes this behaviour, by forcing the circle to match the weight constraint.

The information layer still reports OCCT weights being 1.0, but the visual artifact is fixed.