Path: Fix feature selection and handling
Fix error when working with selections containing edges and faces.  This fix effectively processes edges and only passes faces on to next processing step.  This simplifies the code a bit and changes some `obj.Base` references to `otherBase` to improve readability.
Path: Improve `obj.Base` reading and processing

`obj.Base` is first read in `self._processEdges()`.  All edges are identified and processed there.  Remaining features are saved in `remainingObjBaseFeatures` to be processed in the next section of the Profile code.
Path: Change `if` to `elif`