Start: Properly hash thumbnail filenames Image filenames should be encoded as URI before being hashed.
Also assume that filenames are already utf8 because the community
has been advertiseing utf8 usage since the beginning of this centry.

Calling .encode("utf8") on strings that are already in utf8 simply
raises the following exception:

  UnicodeDecodeError: 'ascii' codec can't decode byte ...
for non-ascii (already utf8) strings.

It's in fact impossible to precisely determine pathname encoding
because different components within the path may have different
encodings, e.g., a utf8 directory name followed by an MBCS filename
is valid on Linux native filesystems.
It's the user's responsibility to keep the iocharset consistent.