Path: Restore removal shape envelope buffer
This buffer application was removed in PR #4978 with removal of rotational code.

The buffer is required to provided adequate significant digit or rounding  in later processing of envelope in Path.Area().

NOTE:  I suggest separate `removalshape` envelopes  be created without the necessary `buffer` adjustment to allow for usage of the `removalshape` for subsequent REST type operations.