[FEM] Add modal add/remove for dispacement constraint
With this, add/remove buttons for displacement constraint will put the user in a
selection mode if pressed when nothing is selected.

Superclass `TaskFemConstraintOnBoundary` Allows code reuse for other
constraints. We do not need this in `TaskFemConstraints` because it is only used
in a subset of constraints, particularly that applies on only some of the
boundary entities.

TODO: Confirm Undo/Redo works appropriately.
TODO: Optimize `onSelectionChanged()`.