[Sketcher] Workaround for Elements theme issue
After merging the change to the Constraint status label in 5fd0c49b7,
the TaskSketcherElements would sometimes be reduced in size to just a
line or two, if multiple task views were expanded and a stylesheet was
enabled. This commit introduces a minimum size to the TSE to prevent
that from occurring.